requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
wikipedia==1.4.0
urllib3==2.2.3
